Setup component folder structure for UI modularization
- Created directory structure for modular components:
  - /src/components/library/ for library-specific components
  - /src/components/common/ for shared UI components
  - /src/components/reservations/ for reservation-related components
- Moved existing components to appropriate folders
- Updated import paths to maintain functionality

Closes #82
