+++ frontend/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/api/axiosInstance.ts","./src/api/libraryService.ts","./src/components/LibraryBrowser.tsx","./src/components/PrivateRoute.tsx","./src/components/RecentReservations.tsx","./src/components/ReservationForm.tsx","./src/contexts/AuthProvider.tsx","./src/contexts/LibraryContext.tsx","./src/pages/Login.tsx","./src/pages/Logout.tsx","./src/pages/MyReservations.tsx","./src/pages/ReservationPage.tsx","./src/pages/Signup.tsx"],"version":"5.7.3"}
+{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/api/axiosInstance.ts","./src/api/libraryService.ts","./src/components/LibraryBrowser.tsx","./src/components/PrivateRoute.tsx","./src/components/reservations/RecentReservations.tsx","./src/components/reservations/ReservationForm.tsx","./src/contexts/AuthProvider.tsx","./src/contexts/LibraryContext.tsx","./src/pages/Login.tsx","./src/pages/Logout.tsx","./src/pages/MyReservations.tsx","./src/pages/ReservationPage.tsx","./src/pages/Signup.tsx"],"version":"5.7.3"}
